wip
- change enum Board_type to uint32_t
- Schneider LXM32i Esc
- zeyu_hand esc id
- add CoE startup

+++ include/iit/ecat/advr/imu_Vn_esc.h
@@ -207,7 +207,7 @@ public:
     virtual const objd_t * get_SDOs() {
         return SDOs;
     }
-    virtual uint16_t get_ESC_type() {
+    virtual uint32_t get_ESC_type() {
         return IMU_VECTORNAV;
     }
